[BREAKING] Migrate to hikari-cp from c3p0
You'll need to change the way you init your system (the constructor
has changed slightly, including the name) and you probably want to
pass in your system-registry, so that you get metrics about pool
usage. We also need to change the places that we broke the component
abstraction by directly accessing :connection prior to passing to
clojure.java.jdbc functions. Now we should be able to pass the db
component directly (for instance in dao/jdbc-runner). See analytics
for an example of how to do it.